service/back/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
qdrant-client
python-dotenv
pydantic-settings
